+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog
+H	# Changelog
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
@@ -58,6 +58,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+ * WGS (World Geodetic System) functions moved to EarthConstants module. Other Earth constants simplified, renamed (#524 and (!522).
  * Keep `ID_DATA` when the `dia*` files are grouped with `midas.reunir_obs_mpi` (#552 and !521)
    * With `dumpToSqliteAfterThinning=.true.` in `NAMBIASSAT`, the output files have the prefix `bcr` and are grouped together in `midas.reunir_obs_mpi`.
  * Improve the documentation on how to start a branch associated to an